requirements.txt
asgiref==3.8.1
Django==5.1
django-cors-headers==4.4.0
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
Markdown==3.6
pillow==10.4.0
PyJWT==2.9.0
python-dotenv==1.0.1
shortuuid==1.0.13
sqlparse==0.5.1
typing_extensions==4.12.2
